build: rename test util on cw-sdk
Ref. metriport/metriport#846
